Replace problematic docstring with comment
This is a syntax problem which is causing a `SyntaxError` if this file is run (unchanged) using `pox.py log.level --DEBUG misc.videoSlice`.

Basically, the docstring is being interpreted as a string that is a key for the `self.portmap` dict without a corresponding value. With my change, it works fine.